fix logging for named scopes
